design: 공유 확장뷰 리팩토링 및 디자인 수정
