+++ APEX/StashShare/Resources/Assets.xcassets/Color/KeyColor/Primary/Contents.json
@@ -1,0 +1,6 @@
+{
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  }
+}
