+++ APEX/StashShare/Resources/Assets.xcassets/Color/KeyColor/Secondary/Contents.json
@@ -1,0 +1,6 @@
+{
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  }
+}
