+++ APEX/StashShare/SubView/AttachBarExt.swift
@@ -1,28 +1,16 @@
 //
-//  AttachBar.swift
-//  APEX
+//  AttachBarExt.swift
+//  StashShare
 //
-//  Created by AI Assistant on 10/29/25.
+//  Extracted from ShareSheetView for reuse and organization.
 //
 
 import SwiftUI
 
-struct ShareAttachmentItem: Identifiable, Equatable {
-    enum Kind: Equatable {
-        case image(UIImage)
-        case video(URL?, thumbnail: UIImage?)
-        case file(URL)
-        case text(String)
-        case audio(URL)
-    }
-    let id: UUID
-    var kind: Kind
-}
-
-struct AttachBar: View {
+struct AttachBarExt: View {
     let items: [ShareAttachmentItem]
     var onRemove: (ShareAttachmentItem) -> Void
-
+    
     private enum Metrics {
         static let itemSize: CGFloat = 72
         static let corner: CGFloat = 3.95
@@ -30,11 +18,10 @@ struct AttachBar: View {
         static let xSize: CGFloat = 16
         static let xTapSize: CGFloat = 28
     }
-
+    
     var body: some View {
         ScrollView(.horizontal, showsIndicators: false) {
             HStack(spacing: Metrics.spacing) {
-                // Only show visual media (image/video) in the attach bar
                 ForEach(items.filter { item in
                     switch item.kind {
                     case .image, .video:
@@ -50,14 +37,14 @@ struct AttachBar: View {
             .padding(.vertical, 4)
         }
     }
-
+    
     @ViewBuilder
     private func itemView(_ item: ShareAttachmentItem) -> some View {
         ZStack(alignment: .topTrailing) {
             content(for: item)
                 .frame(width: Metrics.itemSize, height: Metrics.itemSize)
                 .clipShape(RoundedRectangle(cornerRadius: Metrics.corner, style: .continuous))
-
+            
             Button {
                 onRemove(item)
             } label: {
@@ -77,7 +64,7 @@ struct AttachBar: View {
         .background(Color("BackgroundSecondary"))
         .clipShape(RoundedRectangle(cornerRadius: Metrics.corner, style: .continuous))
     }
-
+    
     @ViewBuilder
     private func content(for item: ShareAttachmentItem) -> some View {
         switch item.kind {
@@ -99,23 +86,10 @@ struct AttachBar: View {
                     .foregroundColor(.white.opacity(0.95))
                     .shadow(radius: 2)
             }
-        case .file:
-            // Non-visual in AttachBar (hidden via filter above)
-            Color.clear
-        case .text:
-            // Non-visual in AttachBar (hidden via filter above)
-            Color.clear
-        case .audio:
-            // Non-visual in AttachBar (hidden via filter above)
+        default:
             Color.clear
         }
     }
 }
 
-#Preview {
-    let sample = [
-        ShareAttachmentItem(id: UUID(), kind: .image(UIImage(systemName: "person.crop.circle")!)),
-        ShareAttachmentItem(id: UUID(), kind: .video(nil, thumbnail: nil))
-    ]
-    return AttachBar(items: sample, onRemove: { _ in })
-}
+
